Security config

+++ v_story_be/src/main/java/com/v_story/v_story_be/Interfaces/RepositoryInterface.java
@@ -5,6 +5,6 @@ import java.util.Optional;
 
 public interface RepositoryInterface<T> {
     public List<T> getAll();
-    public Optional<T> get(Integer id);
+    public Optional<T> get(Long id);
     public Optional<T> save(T t);
 }
